requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2021.4.13=h06a4308_1
certifi=2020.12.5=py36h06a4308_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
dataclasses=0.8=pyh4f3eec9_6
datasets=1.6.2=pypi_0
dill=0.3.3=pypi_0
filelock=3.0.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.4.0=pypi_0
huggingface-hub=0.0.8=pypi_0
idna=2.10=pypi_0
importlib-metadata=4.0.1=pypi_0
intel-openmp=2021.2.0=h06a4308_610
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
multiprocess=0.70.11.1=pypi_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=hff7bd54_1
numpy=1.19.5=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pypi_0
pandas=1.1.5=pypi_0
pillow=8.2.0=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
pyarrow=4.0.0=pypi_0
pyparsing=2.4.7=pypi_0
python=3.6.13=hdb3f193_0
python-dateutil=2.8.1=pypi_0
pytorch=1.7.1=py3.6_cuda10.1.243_cudnn7.6.3_0
pytz=2021.1=pypi_0
readline=8.1=h27cfd23_0
regex=2021.4.4=pypi_0
requests=2.25.1=pypi_0
sacremoses=0.0.45=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=py36h06a4308_0
sklearn=0.0=pypi_0
sqlite=3.35.4=hdfb4753_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.2=pypi_0
torchaudio=0.7.2=py36
torchvision=0.8.2=py36_cu101
tqdm=4.49.0=pypi_0
transformers=4.5.1=pypi_0
typing-extensions=3.10.0.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xxhash=2.0.2=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
